Move the feature flags to provider
