+++ src/page-multi-facility/MultiFacilityImpactDashboard.tsx
@@ -10,13 +10,13 @@ import Colors from "../design-system/Colors";
 import iconAddSrc from "../design-system/icons/ic_add.svg";
 import Loading from "../design-system/Loading";
 import TextLabel from "../design-system/TextLabel";
-import { useFlag } from "../feature-flags";
 import useFacilitiesRtData from "../hooks/useFacilitiesRtData";
 import { EpidemicModelProvider } from "../impact-dashboard/EpidemicModelContext";
 import {
   getFacilitiesRtDataById,
   updateFacilityRtData,
 } from "../infection-model/rt";
+import { useFlag } from "../providers/feature-flags";
 import { FacilityContext } from "./FacilityContext";
 import FacilityRow from "./FacilityRow";
 import FacilityRowPlaceholder from "./FacilityRowPlaceholder";
